Hotfix ETP-4616: Fixed comment sonnar
